v14: Fix margin inconsistency with framework version of PreferenceCategory
The preference_category_material.xml layout files appear to have been
originally copied from AOSP's frameworks/base and thus contains:

    android:layout_marginBottom="16dip"

This is fine in the AOSP framework because its PreferenceFragment uses a
ListView, which ignores layout_marginBottom and uses a standard 1dip gap
between items.

However, preference-v14's PreferenceFragment uses a RecyclerView, which
respects the bottom margin. As a result, each preference category
TextView has a large gap below it. This commit updates the layout to use
a 1dip bottom margin instead so that it's visually identical to the AOSP
layout.

Change-Id: I600ee3db0314236356880071befffbbd46cba59f
Signed-off-by: mydongistiny <jaysonedson@gmail.com>